Feat/validacoes

+++ src/main/java/br/com/rosario/desafio_checkpoint_java_nivel1/service/SalaService.java
@@ -4,12 +4,13 @@ import br.com.rosario.desafio_checkpoint_java_nivel1.dto.AtualizacaoSalaDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.dto.CadastroSalaDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.dto.SalaDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.entity.Sala;
+import br.com.rosario.desafio_checkpoint_java_nivel1.entity.StatusSala;
+import br.com.rosario.desafio_checkpoint_java_nivel1.exception.ValidacaoException;
 import br.com.rosario.desafio_checkpoint_java_nivel1.repository.SalaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class SalaService {
@@ -25,31 +26,51 @@ public class SalaService {
                 .toList();
     }
 
-    public Optional<SalaDTO> buscarPorID(Long id) {
+    public SalaDTO buscarPorID(Long id) {
         return repository.findById(id)
-                .map(SalaDTO::new);
+                .map(SalaDTO::new)
+                .orElseThrow(() -> new ValidacaoException("Sala não encontrada!"));
     }
 
     public void cadastrarSala(CadastroSalaDTO dto) {
+        if (repository.existsByName(dto.nome())) {
+            throw new ValidacaoException("Nome já cadastrado!");
+        }
         repository.save(new Sala(dto));
     }
 
     public void atualizarSala(AtualizacaoSalaDTO dto) {
-        Sala sala = repository.getReferenceById(dto.id());
+        if (repository.existsByNameAndIdNot(dto.nome(), dto.id())) {
+            throw new ValidacaoException("Nome já cadastrado!");
+        }
+        Sala sala = repository.findById(dto.id())
+                .orElseThrow(() -> new ValidacaoException("Sala não encontrada!"));
         sala.atualizarDados(dto);
     }
 
     public void ativarSala(Long id) {
-        Sala sala = repository.getReferenceById(id);
+        if (repository.existsByIdAndStatusSala(id, StatusSala.DISPONIVEL)) {
+            throw new ValidacaoException("Sala já está ativada!");
+        }
+        Sala sala = repository.findById(id)
+                .orElseThrow(() -> new ValidacaoException("Sala não encontrada!"));
         sala.ativarSala();
     }
 
     public void desativarSala(Long id) {
-        Sala sala = repository.getReferenceById(id);
+        if (repository.existsByIdAndStatusSala(id, StatusSala.INATIVA)) {
+            throw new ValidacaoException("Sala já está desativada!");
+        }
+        Sala sala = repository.findById(id)
+                .orElseThrow(() -> new ValidacaoException("Sala não encontrada!"));
         sala.desativarSala();
     }
 
     public void deletarSala(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ValidacaoException("Sala não encontrada!");
+        }
         repository.deleteById(id);
+
     }
 }
