+++ src/main/java/br/com/rosario/desafio_checkpoint_java_nivel1/service/UsuarioService.java
@@ -4,12 +4,12 @@ import br.com.rosario.desafio_checkpoint_java_nivel1.dto.AtualizacaoUsuarioDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.dto.CadastroUsuarioDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.dto.UsuarioDTO;
 import br.com.rosario.desafio_checkpoint_java_nivel1.entity.Usuario;
+import br.com.rosario.desafio_checkpoint_java_nivel1.exception.ValidacaoException;
 import br.com.rosario.desafio_checkpoint_java_nivel1.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UsuarioService {
@@ -25,22 +25,33 @@ public class UsuarioService {
                 .toList();
     }
 
-    public Optional<UsuarioDTO> buscarPorId(Long id) {
-        return repository
-                .findById(id)
-                .map(UsuarioDTO::new);
+    public UsuarioDTO buscarPorId(Long id) {
+            return repository
+                    .findById(id)
+                    .map(UsuarioDTO::new)
+                    .orElseThrow(() -> new ValidacaoException("Usuário não encontrado!"));
     }
 
     public void cadastrarUsuario(CadastroUsuarioDTO dto) {
+        if (repository.existsByTelefone(dto.telefone())) {
+            throw new ValidacaoException("Telefone já cadastrado!");
+        }
         repository.save(new Usuario(dto));
     }
 
     public void atualizarUsuario(AtualizacaoUsuarioDTO dto) {
-        Usuario usuario = repository.getReferenceById(dto.id());
+        if (repository.existsByTelefone(dto.telefone())) {
+            throw new ValidacaoException("Telefone já cadastrado!");
+        }
+        Usuario usuario = repository.findById(dto.id())
+                        .orElseThrow(() -> new ValidacaoException("Usuário não encontrado!"));
         usuario.atualizarDados(dto);
     }
 
     public void excluirUsuario(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ValidacaoException("Usuário não encontrado!");
+        }
         repository.deleteById(id);
     }
 }
